CT-427: Generalize `XlmKeyDeriver` for usage by other ed25519 coins
